+++ aftiktuna-rust/CHANGELOG.md
@@ -9,10 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
+- Idle animation for aftiks
 - Model format changes:
   - "fixed" layer property replaced by "fixed_orientation" model property
   - "y_offset" layer property replaced by "offset" layer property with both x- and y-offsets
-  - New layer properties for rotation and rotation-anchor
+  - New layer properties for rotation animation and rotation-anchor
 
 ### Contributors for this release